feat(sidebar): add custom hooks for managing space dialogs and optimize space menu item rendering
